analyzer.py
import anthropic
import os
from dotenv import load_dotenv

load_dotenv()

SYSTEM_PROMPT = """Tu es un expert en biologie médicale. Tu analyses des résultats de bilan biologique et tu rédiges un rapport d'interprétation structuré à destination du médecin prescripteur.

Ton rapport doit toujours contenir ces 4 sections :

## 1. Résumé des anomalies
Liste uniquement les valeurs hors normes avec leur degré de gravité (légère / modérée / critique).

## 2. Interprétation clinique
Explique ce que ces résultats signifient cliniquement de façon claire et concise.

## 3. Diagnostics à évoquer
Liste les hypothèses diagnostiques principales à considérer selon le tableau biologique.

## 4. Recommandations
Propose les examens complémentaires ou actions à envisager (contrôle, bilan complémentaire, avis spécialisé...).

Règles :
- Langue : français médical clair
- Si les valeurs sont toutes normales, indique-le explicitement
- Ne pose jamais de diagnostic certain, seulement des hypothèses
- Toujours rappeler que l'interprétation finale appartient au médecin
"""


def analyze(raw_text: str) -> str:
    client = anthropic.Anthropic(api_key=os.environ["ANTHROPIC_API_KEY"])
    message = client.messages.create(
        model="claude-opus-4-7",
        max_tokens=1500,
        system=SYSTEM_PROMPT,
        messages=[
            {
                "role": "user",
                "content": f"Voici les résultats du bilan biologique à analyser :\n\n{raw_text}",
            }
        ],
    )
    return message.content[0].text
